solution/requirements.txt

langgraph>=0.2.50
langchain>=0.3.0
langchain-core>=0.3.0

langchain-groq>=0.2.0
langchain-google-genai>=2.0.0

anyio>=4.0.0
httpx>=0.27.0


python-dotenv>=1.0.0

jsonschema>=4.0.0
pydantic>=2.0.0

colorama>=0.4.0



